Add support for margins and borders on text layers
MARGINS

A new `margins` property is now supported for text layers. This enables
overriding the existing defaults of 4 (horiztonal) and 1 (vertical) to
render padded boxes around the supplied text. The `margins` property
supports the following sub-properties:

* h: horizontal margin
* v: vertical margin
* top: top margin; overrides `v` if both are given
* bottom: bottom margin; overrides `v` if both are given
* left: left margin; overrides `h` if both are given
* right: right margin; overrides `h` if both are given

If the `margin` property is omitted, the behavior is unchanged.

BORDERS

New `border` and `borderWeight` properties are now supported for text
layers. If neither are provided, the behavior is unchanged.

* `border`: A complement to `background` this expects a Panels.Color value
* `borderWidth`: The width of the border. Defaults to 1.

A border is only drawn if `border` is provided (setting `borderWidth` alone
has no effect). There are some design choices here that could go in other
directions. For example, there could be a `border` property that expects
`color` and `width` sub-proerties. It would also be possible to automatically
draw the border in a contrasting color with respect to the background, but
that would preclude the ability to draw a border without a background.
Not sure the best API here.