Give each panel a number
This sets the `number` property of each panel to its monotonically
increasing "page" number according to its relative position in the
list of all sequences. While this makes more sense for truly sequential
content, it can still ensure a unique number is assigned to each even
when creating a branching choose-your-own-adventure-style story.

To what end? This is primarily intended for use in conjunction with the
new `borderRenderFunction`, to allow showing the page number for each
panel.